authentication_provider/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.0.3
coverage==6.2
Flask==2.0.2
Flask-Caching==1.10.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
tomli==2.0.0
Werkzeug==2.0.2
zipp==3.7.0
python-dateutil==2.8.2
flask_cors==3.0.10
bcrypt==3.2.0
pyjwt==2.3.0
